refactor: Separate concerns + prop-driven tabs
